Edit Groups & Edit Group weitestgehend fertig --> Noch nicht intensiv nach Fehlern gesucht

+++ server/services/user.service.js
@@ -64,16 +64,18 @@ const updateUser = async (userID, updateBody) => {
 		throw new ApiError(httpStatus.BAD_REQUEST, "Username already taken");
 	}
 
-	if (!_.isEqual(user.accessible_groups, updateBody.accessible_groups)) {
-		//Remove Member from Group
-		const removedGroups = user.accessible_groups.filter(group => !updateBody.accessible_groups.includes(group))
-		for (groupId of removedGroups) {
-			await groupService.removeTrainer(groupId, user._id)
-		}
-		//Add Member to Group
-		const newGroups = updateBody.accessible_groups.filter(group => !user.accessible_groups.includes(group))
-		for (groupId of newGroups) {
-			await groupService.addTrainer(groupId, { id: user._id, role: user.roles.includes("trainer") ? "trainer" : "assistant" })
+	if (updateBody.accessible_groups !== undefined) {
+		if (!_.isEqual(user.accessible_groups, updateBody.accessible_groups)) {
+			//Remove Member from Group
+			const removedGroups = user.accessible_groups.filter(group => !updateBody.accessible_groups.includes(group))
+			for (groupId of removedGroups) {
+				await groupService.removeTrainer(groupId, user._id)
+			}
+			//Add Member to Group
+			const newGroups = updateBody.accessible_groups.filter(group => !user.accessible_groups.includes(group))
+			for (groupId of newGroups) {
+				await groupService.addTrainer(groupId, { id: user._id, role: user.roles.includes("trainer") ? "trainer" : "assistant" })
+			}
 		}
 	}
